# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PikPakGo API",
        "description": "PikPakGo Travel Marketplace API - Hotels, Vacation Rentals & More",
        "contact": {
            "name": "PikPakGo Support",
            "email": "reservations@pikpakgo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://pickpackgo.in-sourceit.com/api",
            "description": "Production Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/admin/dashboard/stats": {
            "get": {
                "tags": [
                    "Admin - Dashboard"
                ],
                "summary": "Get dashboard overview stats",
                "operationId": "db10d406073eeeca359824caa4c558af",
                "responses": {
                    "200": {
                        "description": "Overview stats"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/dashboard/revenue": {
            "get": {
                "tags": [
                    "Admin - Dashboard"
                ],
                "summary": "Get revenue statistics over time",
                "operationId": "0e89539a382157b3be1d59e93e0354b4",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "daily|weekly|monthly",
                        "schema": {
                            "type": "string",
                            "example": "monthly"
                        }
                    },
                    {
                        "name": "months",
                        "in": "query",
                        "description": "Number of months (default 6)",
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Revenue chart data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/dashboard/bookings-chart": {
            "get": {
                "tags": [
                    "Admin - Dashboard"
                ],
                "summary": "Get booking status breakdown for chart",
                "operationId": "6023168995ab9a49dc8b79ae2e1287a0",
                "responses": {
                    "200": {
                        "description": "Booking chart data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/dashboard/recent-bookings": {
            "get": {
                "tags": [
                    "Admin - Dashboard"
                ],
                "summary": "Get recent bookings for dashboard feed",
                "operationId": "a2ade8468b2f49ee8703947b2d1e8a40",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recent bookings"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/dashboard/top-properties": {
            "get": {
                "tags": [
                    "Admin - Dashboard"
                ],
                "summary": "Get top properties by bookings and revenue",
                "operationId": "97be3e8c093a0f307bb19780ca70cb97",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top properties"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/dashboard/user-growth": {
            "get": {
                "tags": [
                    "Admin - Dashboard"
                ],
                "summary": "Get user registration growth over time",
                "operationId": "9d1a5188bbe8cadd55fe9bf09961a76d",
                "parameters": [
                    {
                        "name": "months",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User growth data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}